1. 修复了内存越界的bug(问题出在数组预设数量不足造成的) 2. 增强了对巨量音符同时流畅发音的算法 3. 修复了解析大midi文件缓慢的bug

+++ Ventrue/src/core/Synth/Channel.cpp
@@ -253,8 +253,8 @@ namespace ventrue
 
 	void Channel::Clear()
 	{
-		memset(ccValue, 0, sizeof(int) * 128);
-		memset(ccCombValue, 0, sizeof(int) * 128);
+		memset(ccValue, 0, sizeof(float) * 128);
+		memset(ccCombValue, 0, sizeof(float) * 128);
 		memset(ccComputedValue, 0, sizeof(float) * 128);
 
 		ccComputedValue[(int)MidiControllerType::PanMSB] = 0;
